backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.30.1

# Database & ORM
SQLAlchemy==2.0.30
sqlmodel==0.0.22
alembic==1.13.1

# Pydantic
pydantic==2.7.4
pydantic-settings==2.3.3

# Auth & Security
python-jose==3.3.0
bcrypt==4.0.1
passlib==1.7.4
passlib[bcrypt]==1.7.4
email-validator==2.1.1

# ML/Embeddings (for KB search stubs)
sentence-transformers==3.0.1
scikit-learn==1.5.0
numpy==1.26.4

# Monitoring & Observability
sentry-sdk==2.38.0
prometheus-fastapi-instrumentator==7.1.0

# ASGI framework
starlette==0.48.0

gunicorn==22.0.0

